refactor: CartItem에서 공통 UI를 BasicCartItem으로 분리하고, CartItem과 OrderItem 구현
