+++ src/components/Order/OrderContainer/OrderItemList/OrderItem/OrderItem.style.ts
@@ -1,0 +1,13 @@
+import { css } from "@emotion/react";
+
+export const OrderCartItemContainerStyle = css({
+  display: "flex",
+  flexDirection: "column",
+  gap: "12px",
+  marginBottom: "12px",
+});
+
+export const OrderItemQuantityStyle = css({
+  fontSize: "12px",
+  fontWeight: "500",
+});
