+++ src/components/common/BasicCartItem/BasicCartItem.style.ts
@@ -1,19 +1,6 @@
 import { css } from "@emotion/react";
 import { COLOR_PALETTE } from "../../../colorPalette";
 
-export const CartItemDetailControlsStyle = css({
-  display: "flex",
-  justifyContent: "space-between",
-  alignItems: "center",
-});
-
-export const CartItemContainerStyle = css({
-  display: "flex",
-  flexDirection: "column",
-  gap: "12px",
-  marginBottom: "12px",
-});
-
 export const CartItemImageStyle = css({
   width: "112px",
   height: "112px",
@@ -42,19 +29,4 @@ export const CartItemPriceStyle = css({
   textAlign: "left",
 
   marginBottom: "24px",
-});
-
-export const CartItemQuantityContainerStyle = css({
-  display: "flex",
-  gap: "4px",
-});
-
-export const CartItemQuantityStyle = css({
-  textAlign: "center",
-  minWidth: "24px",
-});
-
-export const OrderItemQuantityStyle = css({
-  fontSize: "12px",
-  fontWeight: "500",
 });
